+++ src/config.ts
@@ -50,6 +50,20 @@ const config = {
     zhAnalysis: 3,
     enAnalysis: 6,
   },
+  // 每天可领取次数
+  dailyClaimLimits: {
+    zhSummary: 4,
+    enSummary: 3,
+    zhAnalysis: 2,
+    enAnalysis: 1
+  },
+  // 每次可领取点数
+  dailyClaimAmount: {
+    zhSummary: 1,
+    enSummary: 1,
+    zhAnalysis: 1,
+    enAnalysis: 1
+  },
 };
 
 export default config;